@@ -9,7 +9,7 @@ version: 1.4.9+80
 
 environment:
   flutter: ">=3.3.0 <4.0.0"
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.2 <3.0.0"
 
 dependencies:
   flutter:
